[cheriabi] Update types for as_int and as_uint
As the `as_uint` and `as_int` fields are used to store arbitraty types
they must be able to both store a capability value and preserve the
capability tag.  Thereforem, these types are defined as `u/intptr_t`.

Update the static_assert and casts to match this new type.

Align with changes to platform TraceArguments.